e2e: prompt_nested_folders: removed duplicated methods for data classes
