+++ e2e/src/testData/index.ts
@@ -5,3 +5,4 @@ export * from './conversationHistory/conversationData';
 export * from './conversationHistory/folderBuilder';
 export * from './prompts/promptBuilder';
 export * from './prompts/promptData';
+export * from './folders/folderData';
