+++ e2e/src/testData/prompts/promptData.ts
@@ -3,7 +3,7 @@ import { GeneratorUtil } from '@/e2e/src/utils/generatorUtil';
 import { FolderInterface } from '@/src/types/folder';
 import { Prompt } from '@/src/types/prompt';
 
-import { FolderBuilder } from '@/e2e/src/testData/conversationHistory/folderBuilder';
+import { FolderData } from '@/e2e/src/testData/folders/folderData';
 import { PromptBuilder } from '@/e2e/src/testData/prompts/promptBuilder';
 
 export interface FolderPrompt {
@@ -11,44 +11,25 @@ export interface FolderPrompt {
   folders: FolderInterface;
 }
 
-export class PromptData {
+export class PromptData extends FolderData {
   private promptBuilder: PromptBuilder;
-  private folderBuilder: FolderBuilder;
 
   constructor() {
+    super('prompt');
     this.promptBuilder = new PromptBuilder();
-    this.folderBuilder = new FolderBuilder().withType('prompt');
   }
 
   public resetData() {
     this.promptBuilder = new PromptBuilder();
-    this.folderBuilder = new FolderBuilder().withType('prompt');
+    this.resetFolderData();
   }
 
   public prepareDefaultPrompt() {
     return this.promptBuilder.withName(GeneratorUtil.randomString(10)).build();
   }
 
-  public prepareFolder(name?: string) {
-    return this.folderBuilder
-      .withName(name ?? GeneratorUtil.randomString(7))
-      .build();
-  }
-
   public prepareNestedFolder(nestedLevel: number) {
-    const rootFolder = this.prepareFolder();
-    this.resetData();
-    const foldersHierarchy = [rootFolder];
-    for (let i = 1; i <= nestedLevel; i++) {
-      const nestedFolder = this.folderBuilder
-        .withName(GeneratorUtil.randomString(7))
-        .withType('prompt')
-        .withFolderId(foldersHierarchy[foldersHierarchy.length - 1].id)
-        .build();
-      foldersHierarchy.push(nestedFolder);
-      this.resetData();
-    }
-    return foldersHierarchy;
+    return super.prepareNestedFolder(nestedLevel, 'prompt');
   }
 
   public prepareDefaultPromptInFolder(name?: string): FolderPrompt {
